Force-replace all Zi and ZI with Z7 so that there are no separate pdb.
Zi and ZI cause the real compiler to create separate pdb files for each
project.  Attempting to cache the pdb files is difficult because they
are incrementally updated for each object.  My initial attempt at
caching the pdb file with each object failed with a "Permission denied",
presumably because the real compiler for another object in the same
project already had the pdb file open for writing.  I think the
correct solution would be to cache the pdb file once per project, but
clcache is not currently setup to do per-project caching.

So instead, here I'm replacing any Zi or ZI with Z7, which also
generates debug information, but stores it within the object instead
of in a separate pdb file.