Call the real compiler without the VS_UNICODE_OUTPUT environment var.
Warning! This commit breaks unicode support!

The VS_UNICODE_OUTPUT environment variable tells the real compiler
to handle unicode, in which case it cannot use stdout & stderr.
clcache uses stdout & stderr to get the include files from the compiler.
So if VS_UNICODE_OUTPUT is set, clcache does not get the includes, which
means it does not use their hashes as part of the key, and so changes
to the includes do not cause a cache miss as is necessary.

Also note that the showIncludes flag is necessary for clcache to get the
includes from the real compiler.