Cannot cache an invocation without the showIncludes flag.
clcache uses the hashes of the includes provided by the compiler as
part of the key for a cached object, so that if an include changes the
object will be recompiled (aka, header changed miss).  If the
showIncludes flag is not passed to the real compiler then it will not
provide the includes in its output, and clcache will assume there are
no includes.  If that happens, changes to the includes will not
cause the necessary cache miss, and a stale cached object will be
erroneously used.