Re-enables parsing escaped quotes.
The proper approach would be to count the number of consecutive
backslashes, and only if an even number (including zero) preceded
the quotes should we toggle 'insideQuotes'.
Since I've only seen a maximum of two consecutive backslashes so far,
this code only checks for one or two.  If there are two consecutive
backslashes, the first one is escaping the second, which means that
the quote is not being escaped, and so we should toggle 'insideQuotes.'