Do not strip quotes around arguments.
If we strip the quotes from the arguments and then pass the altered
commandline to the real compiler, we are affecting the compilation.
For example, when building with cmake, this is an expected
macro definition: /D "CMAKE_INTDIR=\"Debug\""
Stripping the quotes would change that to: /D CMAKE_INTDIR="Debug"

I ran into this issue when trying to add precompiled header support,
because I was passing the original commandline to the pch compilation,
but the altered commandline to the compilations that used the pch
result, which caused an error about macro redefinition.